Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.15.5
dnspython==2.4.2
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.3.0
python-dotenv==1.0.0
requests==2.32.5
